A few first functions to add and search for books in the database
Contents:
- 1 go file containing the functions and type definitions
- 1 go file wich do a bunch of test calls to the functions (for testing and to show how they are ment to be used)
- 1 text file with sql commands to create the necessary tables (run by writing source [filepath] in the mysql client)
OBS! The sql-file first removes the tables if they already exist (during testing I wanted to be able to run it several times and remake the tables each time) so DON'T run it if there already is a table in the database whit the same name which have data in it we don't want to lose.
Note: the sql commands assumes the data base have already been created and selected. The go testfile has a part in the beginning with log-in settings for the database that might need to be changed.