front end and webserver parts of login page
- the webserver part calls a function in the application part with identical specification as the one in Viktor's branch. The actuall function in this branch is a dummy function though: none of the changes in the databaseoperations file should need to be merged to the main branch as long as Viktor's changes are merged there instead (I still included it so the code in this branch will run as is)
- when a login is successfull, the cookies UserID, Username, Password, IsSeller and IsAdmin is added. Those cookies will be automatically passed along in the header with any future http requests, and can be reached in the webserver through:
cookie=r.Cookie("cookiename")
(where r is the request)
or get all cookies by:
cookielist=r.Cookies()
The return values will be of the type http.Cookie (or in the latter case a list (slice) of such), which is a struct with a bunch of different fields. To get the value use cookie.value.
If the values are needed on the client side,  document.cookie in javascript gives a list of all cookies (as name=value pairs separated by ;). The followin should give the value of a specific cookie (here named test2. I have not tried this myself yet, just copied an example):
const cookieValue = document.cookie
  .split("; ")
  .find((row) => row.startsWith("test2="))
  ?.split("=")[1];
I was planning to make an example page to test this out but won't have time for that until later this afternoon at the earliest (and this should be needed for the page showing a list of a sellers books that Ida talked about doing).